Rename script after testing, cleanup
