+++ ual_teleop/scripts/key_teleop.py
@@ -1,18 +1,46 @@
-#! /usr/bin/env python
+#!/usr/bin/env python
 
-import curses
-import math
 import rospy
+import curses
+import time
+import math
+import collections
 from curses import textpad
-from enum import Enum
 from uav_abstraction_layer.srv import TakeOff, Land
-from uav_abstraction_layer.msg import State
 from geometry_msgs.msg import TwistStamped, PoseStamped
 
-TeleopState = Enum('TeleopState', 'MAIN VELOCITY_CONTROL POSE_CONTROL')
+class NumericalParam(object):
+    def __init__(self, name, min_value, max_value, value):
+        if min_value > max_value:
+            raise ValueError("Min value {} grater than max {}".format(min_value, max_value))
+        if (value < min_value) or (value > max_value):
+            raise ValueError("Value {} not in range [{}, {}]".format(value, min_value, max_value))
+        self.name = name
+        self.min = min_value
+        self.max = max_value
+        self.val = value
+        self.msg = "Initial value: {}".format(self.val)
+
+    def set(self, value):
+        # TODO: check value is float!
+        if value < self.min:
+            self.val = self.min
+            self.msg = "Value [{}] too low, setting to [{}]".format(value, self.val)
+        elif value > self.max:
+            self.val = self.max
+            self.msg = "Value [{}] too high, setting to [{}]".format(value, self.val)
+        else:
+            self.val = value
+            self.msg = "Value correctly set: {}".format(self.val)
+
+global_params = collections.OrderedDict()
+global_params['z_takeoff'] = NumericalParam('z_takeoff', 0.0, 5.0, 2.0)
+global_params['vel_mul'] = NumericalParam('vel_mul', 0.0, 3.0, 1.0)
+global_params['rate_mul'] = NumericalParam('rate_mul', 0.0, 3.0, 1.0)
+global_params['xyz_mul'] = NumericalParam('xyz_mul', 0.0, 0.3, 0.1)
+global_params['yaw_mul'] = NumericalParam('xyz_mul', 0.0, 0.3, 0.1)
 
 class ConsoleInterface(object):
-
     def __init__(self, stdscr):    
         self.screen = stdscr
         self.screen.nodelay(True)
@@ -30,12 +58,9 @@ class ConsoleInterface(object):
         curses.flushinp()
         return c
 
-    def get_string(self):
+    def get_box(self):
         cmd = self.box.edit().strip()
         return cmd
-
-    # def refresh(self):
-    #     self.screen.refresh()
 
     def set_header(self, msg):
         self.screen.move(self.head_y, 0)
@@ -49,8 +74,15 @@ class ConsoleInterface(object):
         self.screen.addstr(self.foot_y, self.win_x, msg)
         self.screen.refresh()
 
-class KeyAxis(object):
+    def reset_box(self):
+        self.screen.move(self.win_y-1, 0)
+        self.screen.clrtoeol()
+        textpad.rectangle(self.screen, self.win_y-1, self.win_x-1, self.win_y+1, self.win_x+16)
+        self.win.clear()
+        self.screen.move(self.win_y, self.win_x)
+        self.screen.refresh()
 
+class KeyAxis(object):
     def __init__(self, inc_key, dec_key):
         self.value = 0.0
         self.step = 0.09
@@ -82,7 +114,6 @@ class KeyAxis(object):
             self.value = 0.0
 
 class KeyJoystick(object):
-
     def __init__(self):
         self.axis = {}
         self.axis["move_forward"] = KeyAxis(curses.KEY_UP, curses.KEY_DOWN)
@@ -91,112 +122,207 @@ class KeyJoystick(object):
         self.axis["move_yaw"] = KeyAxis(ord('d'), ord('a'))
 
     def press(self, key):
-        for action, axis in self.axis.items():
+        for axis in self.axis.values():
             axis.press(key)
 
     def update(self):
-        for action, axis in self.axis.items():
+        for axis in self.axis.values():
             axis.update()
 
     def __str__(self):
         output = "[ "
-        for key, axis in self.axis.items():
+        for axis in self.axis.values():
             output += "{:.3f}".format(axis.value) + " "
         output += "]"
         return output
 
-class KeyTeleop(object):
+class State(object):
+    def __init__(self):
+        self.outcomes = []
+        self.out_labels = {}
 
+    def add_outcome(self, name):
+        self.outcomes.append(name)
+        self.out_labels[name] = name
+        self.out_labels[name[0]] = name  # TODO: Check if repeated
+
+    def execute(self):
+        raise NotImplementedError()
+
+class MainState(State):
     def __init__(self, console):
+        State.__init__(self)
+        self.console = console
+        self.add_outcome('set_params')
+        self.add_outcome('takeoff')
+        self.add_outcome('land')
+        self.add_outcome('vel')
+        self.add_outcome('pose')
+        self.add_outcome('quit')
+
+    def execute(self):
+        main_msg = '[' + '] ['.join(self.outcomes) + ']'
+        self.console.set_header('Main: {}'.format(main_msg))
+        self.console.set_footer('')
+        self.console.reset_box()
+        while not rospy.is_shutdown():
+            current_cmd = self.console.get_box()
+            if current_cmd in self.out_labels:
+                return self.out_labels[current_cmd]
+            self.console.set_footer('Unknown command [{}]'.format(current_cmd))
+
+def set_param(param, console):
+    console.set_header('[{}] current value: {}. Set a new value:'.format(param.name, param.val))
+    console.set_footer('')
+    console.reset_box()
+    value = console.get_box()
+    try:
+        new_value = float(value)
+        console.reset_box()
+        param.set(new_value)
+        console.set_footer("[{}]: {}".format(param.name, param.msg))
+    except ValueError:
+        console.set_footer("Invalid value [{}]".format(value))
+
+class SetParamsState(State):
+    def __init__(self, console):
+        State.__init__(self)
+        self.add_outcome('quit')
         self.console = console
 
+    def execute(self):
+        params = State()
+        for p in global_params:
+            params.add_outcome(p)
+        self.console.set_footer('')
+        self.console.reset_box()
+        while not rospy.is_shutdown():
+            param_msg = '[' + '] ['.join(params.outcomes) + ']'
+            self.console.set_header('Set param: {} [{}]'.format(param_msg, 'quit'))
+            current_cmd = self.console.get_box()
+            if current_cmd in ('quit', 'q'):
+                return 'quit'
+            if current_cmd in params.out_labels:
+                cmd = params.out_labels[current_cmd]
+                set_param(global_params[cmd], self.console)
+            else :
+                self.console.set_footer('Unknown command [{}]'.format(current_cmd))
+
+class TakeoffState(State):
+    def __init__(self, console):
+        State.__init__(self)
+        self.add_outcome('quit')
+        self.console = console
         take_off_url = 'ual/take_off'
-        land_url     = 'ual/land'
-        set_vel_url  = 'ual/set_velocity'
-        set_pose_url = 'ual/set_pose'
         # rospy.wait_for_service(take_off_url)
-        # rospy.wait_for_service(land_url)
         self.take_off = rospy.ServiceProxy(take_off_url, TakeOff)
-        self.land     = rospy.ServiceProxy(land_url,     Land)
-        self.velocity_pub = rospy.Publisher(set_vel_url,  TwistStamped, queue_size=1)
-        self.pose_pub     = rospy.Publisher(set_pose_url, PoseStamped,  queue_size=1)
-        self.ual_state = State()
+
+    def execute(self):
+        z_takeoff = global_params['z_takeoff'].val
+        self.console.set_header('Taking off at {} [m]...'.format(z_takeoff))
+        self.console.set_footer('')
+        self.console.reset_box()
+        # time.sleep(2)
+        self.take_off(z_takeoff, True)  # TODO: Non-blocking?
+        return 'quit'
+
+class LandState(State):
+    def __init__(self, console):
+        State.__init__(self)
+        self.add_outcome('quit')
+        self.console = console
+        land_url = 'ual/land'
+        # rospy.wait_for_service(land_url)
+        self.land = rospy.ServiceProxy(land_url, Land)
+
+    def execute(self):
+        self.console.set_header('Landing...')
+        self.console.set_footer('')
+        self.console.reset_box()
+        # time.sleep(2)
+        self.land(True)  # TODO: Non-blocking?
+        return 'quit'
+
+class VelState(State):
+    def __init__(self, console):
+        State.__init__(self)
+        self.add_outcome('quit')
+        self.console = console
         self.uav_pose = PoseStamped()
         self.uav_yaw = 0.0
-        self.telop_state = TeleopState.MAIN
-        # self.headless = False
-
-    def state_callback(self, data):
-        self.ual_state = data
+        self.pose_sub = rospy.Subscriber('ual/pose', PoseStamped, self.pose_callback)
+        self.velocity_pub = rospy.Publisher('ual/set_velocity', TwistStamped, queue_size=1)
 
     def pose_callback(self, data):
         self.uav_pose = data
         self.uav_yaw = 2.0 * math.atan2(data.pose.orientation.z, data.pose.orientation.w)
 
-    def run(self):
-        takeoff_cmd      = "takeoff"
-        land_cmd         = "land"
-        vel_control_cmd  = "velocity"
-        pose_control_cmd = "pose"
-        quit_cmd         = "quit"
-        main_msg = "Main: [{}] [{}] [{}] [{}] [{}]".format(takeoff_cmd, land_cmd, vel_control_cmd, pose_control_cmd, quit_cmd)
-        velocity_msg = "Velocity control: [q] to return to main"
-        pose_msg = "Pose control: [q] to return to main"
+    def execute(self):
+        self.console.set_header('Velocity control: [q] to quit to main')
+        self.console.set_footer('')
+        self.console.reset_box()
         joy = KeyJoystick()
-
-        rate = rospy.Rate(10) # [Hz]
+        rate = rospy.Rate(10)  # [Hz]
         while not rospy.is_shutdown():
             rate.sleep()
-            if self.telop_state == TeleopState.MAIN:
-                self.console.set_header(main_msg)
-                command = self.console.get_string()
-                if command == takeoff_cmd:
-                    self.console.set_footer("Taking off...")
-                    self.take_off(2.0, False)  # TODO(franreal): takeoff height? param mode?
-                elif command == land_cmd:
-                    self.console.set_footer("Landing...")
-                    self.land(False)
-                elif command == vel_control_cmd:
-                    self.telop_state = TeleopState.VELOCITY_CONTROL
-                elif command == pose_control_cmd:
-                    self.telop_state = TeleopState.POSE_CONTROL
-                elif command == quit_cmd:
-                    break
-                else:
-                    self.console.set_footer("Unknown command [{}]".format(command))
+            keycode = self.console.get_key()
+            if keycode == ord('q'):
+                return 'quit'
             else:
-                keycode = self.console.get_key()
-                if keycode == ord('q'):
-                    self.telop_state = TeleopState.MAIN
-                    self.console.set_footer("")
-                else:
-                    joy.press(keycode)
-                self.console.set_footer(str(joy))
-
-            if self.telop_state == TeleopState.VELOCITY_CONTROL:
-                self.console.set_header(velocity_msg)
+                joy.press(keycode)
+                self.console.set_footer('axes: ' + str(joy))
+                vel_mul = global_params['vel_mul'].val
+                rate_mul = global_params['rate_mul'].val
                 vel_cmd = TwistStamped()
                 vel_cmd.header.stamp = rospy.Time.now()
                 vel_cmd.header.frame_id = 'map'
-                vx = +joy.axis['move_forward'].value
-                vy = -joy.axis['move_right'].value
-                vz = +joy.axis['move_up'].value
-                yaw_rate = -joy.axis['move_yaw'].value
+                vx = +vel_mul * joy.axis['move_forward'].value
+                vy = -vel_mul * joy.axis['move_right'].value
+                vz = +vel_mul * joy.axis['move_up'].value
+                yaw_rate = -rate_mul * joy.axis['move_yaw'].value
                 vel_cmd.twist.linear.x = (vx*math.cos(self.uav_yaw) - vy*math.sin(self.uav_yaw))
                 vel_cmd.twist.linear.y = (vx*math.sin(self.uav_yaw) + vy*math.cos(self.uav_yaw))
                 vel_cmd.twist.linear.z = vz
                 vel_cmd.twist.angular.z = yaw_rate
                 self.velocity_pub.publish(vel_cmd)
 
-            if self.telop_state == TeleopState.POSE_CONTROL:
-                self.console.set_header(pose_msg)
-                pose_cmd = self.uav_pose
+class PoseState(State):
+    def __init__(self, console):
+        State.__init__(self)
+        self.add_outcome('quit')
+        self.console = console
+        self.uav_pose = PoseStamped()
+        self.uav_yaw = 0.0
+        self.pose_sub = rospy.Subscriber('ual/pose', PoseStamped, self.pose_callback)
+        self.pose_pub = rospy.Publisher('ual/set_pose', PoseStamped, queue_size=1)
+
+    def pose_callback(self, data):
+        self.uav_pose = data
+        self.uav_yaw = 2.0 * math.atan2(data.pose.orientation.z, data.pose.orientation.w)
+
+    def execute(self):
+        self.console.set_header('Pose control: [q] to quit to main')
+        self.console.set_footer('')
+        self.console.reset_box()
+        joy = KeyJoystick()
+        rate = rospy.Rate(10)  # [Hz]
+        pose_cmd = self.uav_pose
+        while not rospy.is_shutdown():
+            rate.sleep()
+            keycode = self.console.get_key()
+            if keycode == ord('q'):
+                return 'quit'
+            else:
+                joy.press(keycode)
+                self.console.set_footer('axes: ' + str(joy))
+                xyz_mul = global_params['xyz_mul'].val
+                yaw_mul = global_params['yaw_mul'].val
                 pose_cmd.header.stamp = rospy.Time.now()
                 # pose_cmd.header.frame_id = 'map'
-                dx = +joy.axis['move_forward'].value
-                dy = -joy.axis['move_right'].value
-                dz = +joy.axis['move_up'].value
-                delta_yaw = -joy.axis['move_yaw'].value
+                dx = +xyz_mul * joy.axis['move_forward'].value
+                dy = -xyz_mul * joy.axis['move_right'].value
+                dz = +xyz_mul * joy.axis['move_up'].value
+                delta_yaw = -yaw_mul * joy.axis['move_yaw'].value
                 pose_cmd.pose.position.x += (dx*math.cos(self.uav_yaw) - dy*math.sin(self.uav_yaw))
                 pose_cmd.pose.position.y += (dx*math.sin(self.uav_yaw) + dy*math.cos(self.uav_yaw))
                 pose_cmd.pose.position.z += dz
@@ -205,15 +331,63 @@ class KeyTeleop(object):
                 pose_cmd.pose.orientation.w = math.cos(0.5*yaw)
                 self.pose_pub.publish(pose_cmd)
 
+class StateMachine(object):
+    def __init__(self):
+        self.states = {}
+        self.transitions = {}
+        self.current_state = ''
+        self.final_state = ''
+
+    def set_current_state(self, state_label):
+        self.current_state = state_label
+
+    def set_final_state(self, state_label):
+        self.final_state = state_label
+
+    def add(self, state_label, state_object, transitions):
+        self.states[state_label] = state_object
+        self.transitions[state_label] = transitions
+
+    def execute(self):
+        while not rospy.is_shutdown():
+            out = self.states[self.current_state].execute()
+            self.current_state = self.transitions[self.current_state][out]
+            if self.final_state == self.current_state:
+                break
+
+
 def main(stdscr):
     rospy.init_node('key_teleop')
-
+    
     console = ConsoleInterface(stdscr)
-    teleop = KeyTeleop(console)
+    sm = StateMachine()
+    sm.add('MAIN', MainState(console),
+            transitions={'set_params':'PARAMS', 
+                         'takeoff':'TAKEOFF', 
+                         'land':'LAND', 
+                         'vel':'VEL', 
+                         'pose':'POSE', 
+                         'quit':'END'})
 
-    rospy.Subscriber('ual/state', State, teleop.state_callback)
-    rospy.Subscriber('ual/pose', PoseStamped, teleop.pose_callback)
-    teleop.run()
+    sm.add('PARAMS', SetParamsState(console), 
+            transitions={'quit':'MAIN'})
+
+    sm.add('TAKEOFF', TakeoffState(console), 
+            transitions={'quit':'MAIN'})
+
+    sm.add('LAND', LandState(console), 
+            transitions={'quit':'MAIN'})
+
+    sm.add('VEL', VelState(console), 
+            transitions={'quit':'MAIN'})
+
+    sm.add('POSE', PoseState(console), 
+            transitions={'quit':'MAIN'})
+
+    sm.set_current_state('MAIN')
+    sm.set_final_state('END')
+    sm.execute()
+
 
 if __name__ == '__main__':
     curses.wrapper(main)
